Move ByteArray to the MessageStorage namespace
