+++ src/NLog.Targets.Syslog/MessageCreation/Rfc3164.cs
@@ -5,6 +5,7 @@ using NLog.Layouts;
 using NLog.Targets.Syslog.Policies;
 using System.Globalization;
 using System.Text;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.MessageCreation
@@ -51,7 +52,6 @@ namespace NLog.Targets.Syslog.MessageCreation
         {
             if (!outputPri)
                 return;
-
             buffer.Append(pri, encoding);
         }
