+++ src/NLog.Targets.Syslog/MessageCreation/Rfc5424.cs
@@ -4,6 +4,7 @@
 using NLog.Layouts;
 using NLog.Targets.Syslog.Policies;
 using System.Globalization;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.MessageCreation
