+++ src/NLog.Targets.Syslog/MessageCreation/SdElement.cs
@@ -4,6 +4,7 @@
 using NLog.Targets.Syslog.Policies;
 using System.Collections.Generic;
 using System.Linq;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.MessageCreation
