+++ src/NLog.Targets.Syslog/MessageCreation/SdId.cs
@@ -2,6 +2,7 @@
 // See the LICENSE file in the project root for more information
 
 using NLog.Layouts;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Policies;
 using NLog.Targets.Syslog.Settings;
