+++ src/NLog.Targets.Syslog/MessageCreation/SdParam.cs
@@ -6,6 +6,7 @@ using NLog.Targets.Syslog.Policies;
 using System.Collections.Generic;
 using System.Linq;
 using NLog.Targets.Syslog.Extensions;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.MessageCreation
