+++ src/NLog.Targets.Syslog/MessageSend/MessageTransmitter.cs
@@ -9,6 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using NLog.Common;
 using NLog.Targets.Syslog.Extensions;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.MessageSend
