+++ src/NLog.Targets.Syslog/MessageSend/Udp.cs
@@ -4,6 +4,7 @@
 using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.MessageSend
