+++ src/NLog.Targets.Syslog/MessageStorage/ByteArray.cs
@@ -5,7 +5,7 @@ using System;
 using System.IO;
 using System.Text;
 
-namespace NLog.Targets.Syslog
+namespace NLog.Targets.Syslog.MessageStorage
 {
     internal class ByteArray : IDisposable
     {
