+++ src/NLog.Targets.Syslog/Policies/TruncateToComputedValuePolicy.cs
@@ -3,6 +3,7 @@
 
 using NLog.Common;
 using NLog.Targets.Syslog.Extensions;
+using NLog.Targets.Syslog.MessageStorage;
 using NLog.Targets.Syslog.Settings;
 
 namespace NLog.Targets.Syslog.Policies
